Dynamic load global content to realize a perf gian
